Cleaning up the files so that the program will build without issue.
