--- NameGeneratorInterface/NameGeneratorInterface/NameGeneratorInterface.csproj.user
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'">
-    <StartArguments>C:\Users\maxfa\source\repos\GeneratorFiles\SerializedLetterData.txt C:\Users\maxfa\source\repos\GeneratorFiles\SerializedFirstLetterData.txt</StartArguments>
-  </PropertyGroup>
-</Project>
